Make content default `author` to `created_by`.
Author is not always the user that creates the model, but if no author
is specified, I think it makes sense to set the user as the author,
if no other author is supplied.

I wasn't sure how to do this without using `null=True` (`if not
self.author` threw an exception otherwise), so thats that.