Use structlog for logging
This makes it easier to parse logs later.
The authentication classes is kind of ugly but it seems to be the easiest way to attach the user to the log context.
This PR also has a simple health package. We need to implement more advanced health checks at a later point.