Remove social groups (#774)
* Move the old descriptionLong into AbakusGroup though a new model

* [NOT TESTED] Attempt at removing the `social_gropus` app.

Fixture generation is needed to test this properly, but this is
currently broken. I'll just push this anyways :)

* Make `/memberships` a nestes route of `groups`.

* Attempt to fix group membership permissions

* Remove `social_groups` yet again, and fix rebase fuckups

* Regenerate group fixtures

I think I messed up the rebase. This is a fresh run, so everything
should be alright now :)

* Remove `AbakusGroup.is_interest_group`, since it's covered by type.

* Handle group permissions correctly for interest groups

* Remove delete group permissions for all users.

* Remove `group_text`.

The motivation behind `group_text` was to avoid fetching the entire text
entire we load a group object, which happens quite often when managing
permissions. It turns out Django has a thing designed to to exactly this
- `defer`, so we can use this instead of messing around with an extra
model all over the place.

* Remove fixture which is no longer present