Add `abakus_group` from the request to membership create
Since we added `abakus_group` in the Membership serializer, it wants us
to supply the group on create as well. Hoewever, this is already in the
URL, since we POST to `/groups/123/memberships` on membership creation.
By pulling `group_pk` from the request, and add it to the data, we don't
have to send the group pk twice (once in the URL and once in the data).