Refresh registration before response
Running celery in eager mode causes the registration response to be
delivered to the client before the websocket with the registration
status. This commit refreshes the registration before returning the
response, meaning that the registration data is correct when it arrives
on the client. This means that the tests will differ when running celery
in eager or not, so they are changed to test against the database values
instead of hardcoded `PENDING`, as it would be when running in production
mode.

Also changes the stripe payment intent status for success to the correct
string.