Allow unauthenticated users list & view permission on companies
This first of all allows proper use of the companyinterest form, where a
user can select their company when filling in the form, without the need
for a user. It also opens up for listing all companies, as well as
viewing their details, without authentication.

This also changes up the permissions on companies a bit. The
CompanyAdminViewset (bdb) now uses the "bdb" keyword instead of
"companies".