Fix race conditions in registration tasks
When two instances of async_register process at the same time, a race
condition causing the registration_status to be set to FAILURE_REGISTER,
while it is actually successfully registered could occur. This should
fix this by locking the registration when performing the registration.
Additionally, this makes the failure handing of the register task more
robust and should not corrupt the registration.

Co-authored-by: Ludvig Hozman <ludvig123@gmail.com>