Add weekly email task
This will add a task for creating and sending an automated weekly email
to each student every week, consisting of events with opening for
registrations, new joblistings and the latest weekly.
It uses send_mass_mail instead of handling the emails using notifications.
This means moving weekly_mail outside of notifications, so we have
control of how the email is sent.