Return consents when user is admitted to event
Also when the user is on waiting list.

This fixes an issue where the consent would not be returned when an
existing registration for the user on the event existed, either in a
pool or on the waiting list. This fixes that by making sure we return
the consent object when this is the case.