Lock pools as well as event when bumping
Previously, only the event itself was locked. When a registration is
registered to an event, we only lock the pool, since we don't actually
care about the event unless it's merged. This means that we can see
race conditions when we lock the _event_ when bumping (e.g. in
`unregister()`), and a single _pool_ in `register()` while actually
modifying the reg count of that same _pool_ in `unregister()`.

By locking the pools in all of these places, as well as the event while
bumping, we make sure that only one of these happens at a time.

This bug has also been hard to spot since the task that checks for
incorrect pool counters used an incorrect query, not including all
events without a merge time (i.e. a single pool). This task is fixed
now, as well as an incorrect fixture exemplifying that this test did not
work.