Fix incorrect waiting list sorting
On some events, usually with many users in the waiting list, the waiting
list is consistently wrong for all users (i.e. the sorting is the same,
incorrect for all users) when admitted to the event.

I'm not 100% sure why this works, but changing the query for the
prefetching on the event retrieve endpoint consistently gives the
correct sorting.