Move setup conf into `pyproject.toml`
`flake8` configuration had to be in seperate `.flake8` file since
it is not supported in `pyproject.toml` ... The `flake8` team even
locked an issue suggesting the feature because it was "too heated" ...
Maybe one day.

The previous `black` configuration never actually worked, so we'll have to configure it properly later.