Introduce "too late" as an event presence status
Similar to the "not present" status, the user will automatically recieve
a penalty. I figured all previous penalties can be removed since only
the newest one should matter.

I would like some input as to how this will affect other penalties,
e.g. given due to late payments (or other things?), as they will also
be removed if the presence is changed to something worthy of a penalty.
Should we consider to introduce "penalty types" stored on the record,
e.g. "presence", "payment", "other"?

Also, this resolves a bug, as it is currently not possible to auto delete
previous penalties related to the event due to the `.delete()` not being
at the end the atomic block. Not sure why it had to be atomic in the
first place.